added log util file, updated policies handler to fail fast on authEntityData, created an error handler for AuthEntityData response failures
